Consolidate modules; cleanup
